Moved functionality post to own component
